rename `components` folder to `features`
