+++ src/App.js
@@ -1,8 +1,8 @@
 import './App.css'
 import { Provider } from 'react-redux'
-import AppHeader from './components/AppHeader'
-import AppFooter from './components/AppFooter'
-import Counter from './components/Counter'
+import AppHeader from './features/AppHeader'
+import AppFooter from './features/AppFooter'
+import Counter from './features/Counter'
 import store from './store'
 
 function App() {
